+++ packages/elements/src/components/ino-markdown-editor/extensions/bullet_list.ts
@@ -1,0 +1,16 @@
+import {wrappingInputRule} from "@tiptap/core";
+import BulletList from "@tiptap/extension-bullet-list";
+
+export const customInputRegex = /^\s*([-+*]\s[^\[])$/
+
+export default BulletList.extend({
+  addInputRules() {
+    return [
+      wrappingInputRule({
+        find: customInputRegex,
+        type: this.type,
+      }),
+    ]
+  },
+});
+