+++ packages/elements/src/components/ino-markdown-editor/markdown-serializer.ts
@@ -1,10 +1,10 @@
 import {
   defaultMarkdownParser,
   defaultMarkdownSerializer,
-  MarkdownParser,
   MarkdownSerializer,
 } from 'prosemirror-markdown';
-import { Node as ProsemirrorNode, Schema } from 'prosemirror-model';
+import { Node as ProsemirrorNode } from 'prosemirror-model';
+import taskLists from '@hedgedoc/markdown-it-task-lists'
 import Bold from '@tiptap/extension-bold';
 import Code from '@tiptap/extension-code';
 import Italic from '@tiptap/extension-italic';
@@ -14,11 +14,13 @@ import Heading from '@tiptap/extension-heading';
 import HorizontalRule from '@tiptap/extension-horizontal-rule';
 import Paragraph from '@tiptap/extension-paragraph';
 import OrderedList from '@tiptap/extension-ordered-list';
-import BulletList from '@tiptap/extension-bullet-list';
+import BulletList from './extensions/bullet_list';
 import ListItem from '@tiptap/extension-list-item';
 import CodeBlock from '@tiptap/extension-code-block';
 import HardBreak from '@tiptap/extension-hard-break';
 import Link from '@tiptap/extension-link';
+import TaskItem from "./extensions/task_item";
+import TaskList from "./extensions/task_list";
 
 // Source is partially taken from
 // https://gitlab.com/gitlab-org/gitlab/-/blob/master/app/assets/javascripts/content_editor/services/markdown_serializer.js
@@ -64,47 +66,15 @@ const defaultSerializerConfig = {
     [ListItem.name]: defaultMarkdownSerializer.nodes.list_item,
     [Paragraph.name]: defaultMarkdownSerializer.nodes.paragraph,
     [Text.name]: defaultMarkdownSerializer.nodes.text,
+    [TaskItem.name]: (state, node) => {
+      state.write(`[${node.attrs.checked ? 'x' : ' '}] `);
+      state.renderContent(node);
+    },
+    [TaskList.name]: (state, node) => {
+      if (node.attrs.numeric) defaultMarkdownSerializer.nodes.ordered_list(state, node);
+      else defaultMarkdownSerializer.nodes.bullet_list(state, node);
+    },
   },
-};
-
-// Source taken from
-// prosemirror-markdown/src/from_markdown.js
-const {
-  bullet_list,
-  ordered_list,
-  link,
-  heading,
-  fence,
-}: MarkdownParser["tokens"] = defaultMarkdownParser.tokens;
-
-const defaultParserTokens: MarkdownParser["tokens"] = {
-  blockquote: { block: Blockquote.name },
-  paragraph: { block: Paragraph.name },
-  list_item: { block: ListItem.name },
-  bullet_list: {
-    ...bullet_list,
-    block: BulletList.name,
-  },
-  ordered_list: {
-    ...ordered_list,
-    block: OrderedList.name,
-  },
-  heading,
-  code_block: { block: CodeBlock.name, noCloseToken: true },
-  fence: {
-    ...fence,
-    block: CodeBlock.name,
-  },
-  hr: { node: HorizontalRule.name },
-  hardbreak: { node: HardBreak.name },
-
-  [Italic.name]: { mark: Italic.name },
-  em: { mark: Italic.name },
-  [Bold.name]: { mark: Bold.name },
-  strong: { mark: Bold.name },
-  link,
-  code_inline: { mark: Code.name, noCloseToken: true },
-  s: { mark: Strike.name },
 };
 
 const markdownSerializer = new MarkdownSerializer(
@@ -112,21 +82,19 @@ const markdownSerializer = new MarkdownSerializer(
   defaultSerializerConfig.marks
 );
 
-let markdownParser = null;
+let markdownRenderer = null;
 
 export default {
   serialize: (content: ProsemirrorNode): string =>
     markdownSerializer.serialize(content, { tightLists: true }),
-  parse: (markdownText: string, schema: Schema): ProsemirrorNode => {
-    if (!markdownParser) {
-      markdownParser = new MarkdownParser(
-        schema,
-        // list of markdown-it rules: https://github.com/markdown-it/markdown-it/issues/289
-        defaultMarkdownParser['tokenizer'].enable('strikethrough'),
-        defaultParserTokens
-      );
+  parse: (markdownText: string): string => {
+    if (!markdownRenderer) {
+      // list of markdown-it rules: https://github.com/markdown-it/markdown-it/issues/289
+      markdownRenderer = defaultMarkdownParser['tokenizer']
+        .enable('strikethrough')
+        .use(taskLists, { enabled: true, label: true, labelAfter: true });
     }
 
-    return markdownParser.parse(markdownText);
+    return markdownRenderer.render(markdownText);
   },
 };