+++ packages/storybook/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [7.1.0](https://github.com/inovex/elements/compare/v7.0.0...v7.1.0) (2022-07-18)
+
+
+### Features
+
+* **elements|ino-carousel:** add `valueChange` event to inform about slide changes on arrow click ([#650](https://github.com/inovex/elements/issues/650)) ([6f98a31](https://github.com/inovex/elements/commit/6f98a317294c9789d5fed72e9210d5f2adf5dacf))
+* **elements|ino-icon-button:** add support for custom icons via default slot ([#600](https://github.com/inovex/elements/issues/600)) ([e2dae7d](https://github.com/inovex/elements/commit/e2dae7dd8b03664d07c7535f573b33cd4075f933))
+* **elements|ino-markdown-editor:** support task lists ([#660](https://github.com/inovex/elements/issues/660)) ([47a71c2](https://github.com/inovex/elements/commit/47a71c25faea9042987706badb5f00e36d021d53))
+* **elements|ino-radio-group:** add option to align radios vertically & horizontally ([#652](https://github.com/inovex/elements/issues/652)) ([2106dfd](https://github.com/inovex/elements/commit/2106dfdd85abd0a93b3325eb82216bddd1e4ea2b))
+* **elements|ino-radio-group:** support keyboard navigation ([#644](https://github.com/inovex/elements/issues/644)) ([b0bf72f](https://github.com/inovex/elements/commit/b0bf72f860fed2bb64d3d7c6589737a01bb59f2b))
+* **elements|ino-range:** support input of intervals ([#605](https://github.com/inovex/elements/issues/605)) ([2f7d8fd](https://github.com/inovex/elements/commit/2f7d8fdb42fab55402adc8913502046a4e043453))
+
+
+### Miscellaneous
+
+* **storybook:** update storybook to 6.5 ([#648](https://github.com/inovex/elements/issues/648)) ([a9ed814](https://github.com/inovex/elements/commit/a9ed814aa1434614492c1b406fab91d38c51da74))
+
 ## [7.0.0](https://github.com/inovex/elements/compare/v6.2.0...v7.0.0) (2022-05-23)